Add rest module and move error handling
