+++ chat-server/src/main/java/com/cyster/flux/chat/ChatController.java
@@ -5,6 +5,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import com.cyster.rest.ErrorResponse;
+import com.cyster.rest.RestException;
 import org.springframework.ai.chat.client.ChatClient;
 import org.springframework.ai.chat.model.ChatModel;
 import org.springframework.ai.tool.ToolCallbackProvider;
