+++ chat-server/src/main/java/com/cyster/flux/chat/ChatServer.java
@@ -3,6 +3,7 @@ package com.cyster.flux.chat;
 import com.cyster.weather.WeatherServicePackageMarker;
 import com.cyster.weather.tool.WeatherToolPackageMarker;
 import com.cyster.weather.tool.WeatherTools;
+import com.cyster.rest.RestPackageMarker;
 import org.springframework.ai.tool.ToolCallbackProvider;
 import org.springframework.ai.tool.method.MethodToolCallbackProvider;
 import org.springframework.boot.SpringApplication;
@@ -13,7 +14,8 @@ import org.springframework.context.annotation.Bean;
     scanBasePackageClasses = {
       ChatServer.class,
       WeatherServicePackageMarker.class,
-      WeatherToolPackageMarker.class
+      WeatherToolPackageMarker.class,
+      RestPackageMarker.class
     })
 public class ChatServer {
   public static void main(String[] args) {
