+++ rest/src/main/java/com/cyster/rest/RestExceptionDocumentationConfig.java
@@ -1,4 +1,4 @@
-package com.cyster.flux.chat;
+package com.cyster.rest;
 
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.media.Content;
