+++ rest/src/main/java/com/cyster/rest/RestPackageMarker.java
@@ -1,0 +1,6 @@
+package com.cyster.rest;
+
+/** Marker class for component scanning of the rest module. */
+public final class RestPackageMarker {
+  private RestPackageMarker() {}
+}
